+++ Sources/Nuke/Caching/Cache.swift
@@ -53,7 +53,9 @@ final class Cache<Key: Hashable, Value>: @unchecked Sendable {
         self.memoryPressure.resume()
 
 #if os(iOS) || os(tvOS)
-        self.registerForEnterBackground()
+        Task {
+            await self.registerForEnterBackground()
+        }
 #endif
 
 #if TRACK_ALLOCATIONS
@@ -70,7 +72,7 @@ final class Cache<Key: Hashable, Value>: @unchecked Sendable {
     }
 
 #if os(iOS) || os(tvOS)
-    private func registerForEnterBackground() {
+    @MainActor private func registerForEnterBackground() {
         notificationObserver = NotificationCenter.default.addObserver(forName: UIApplication.didEnterBackgroundNotification, object: nil, queue: nil) { [weak self] _ in
             self?.clearCacheOnEnterBackground()
         }